# packages/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Voice Chat API
  version: 0.1.0
  description: |
    Branching voice/text conversations in public and private rooms.

    Identity is guest-first: clients call `POST /auth/guest` on first visit to get a
    session cookie, and may later upgrade that same account with `POST /auth/register`.

    The server stores conversation facts only (`parentId`); branches are derived
    client-side. Deleting a capture removes its text, media, and stored file
    everywhere it was shared. The placement stays as a tombstone (`deletedAt` set)
    so replies keep their parent.
  license:
    name: UNLICENSED
    identifier: UNLICENSED
servers:
  - url: http://localhost:3001
    description: Local dev
security:
  - bearerAuth: []
...
paths:
...
  /users/me:
    patch:
      operationId: updateCurrentUser
      summary: Update the current user
      tags: [users]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateMeInput' }
      responses:
        '200':
          description: Updated user
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /users/me/mutes:
    get:
      operationId: listMutes
      summary: People (and bots) I have muted — personal and global (doc/08 I5)
      tags: [users]
      responses:
        '200':
          description: Muted users
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthorList' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /users/me/mutes/{userId}:
    parameters:
      - $ref: '#/components/parameters/UserId'
    put:
      operationId: muteUser
      summary: Mute someone everywhere. Their items reach you content-less; others are unaffected. Idempotent.
      tags: [users]
      responses:
        '200':
          description: Muted
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthorResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      operationId: unmuteUser
      summary: Unmute. Idempotent.
      tags: [users]
      responses:
        '200':
          description: Unmuted
          content:
            application/json:
              schema: { $ref: '#/components/schemas/NullResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /users/me/avatar:
    post:
      operationId: uploadAvatar
      x-fastify-config:
        rateLimit: { max: 20, timeWindow: '1 hour' }
      summary: Replace the current user's avatar
      description: |
        Signed-in accounts only. Stores the image and points the profile at it.
        The previous avatar file is deleted. Guests receive 403. JPEG, PNG, GIF, or WebP.
      tags: [users]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema: { $ref: '#/components/schemas/UploadAvatarInput' }
      responses:
        '200':
          description: Updated user
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '413':
          description: File too large
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '415':
          description: Unsupported file type
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '429': { $ref: '#/components/responses/TooManyRequests' }


  # ─── rooms ─────────────────────────────────────────────────────────────────
  /rooms:
    get:
      operationId: listRooms
      summary: Browse public rooms, most recently active first
      tags: [rooms]
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/SearchQuery'
        - in: query
          name: topic
          schema: { type: string, minLength: 1, maxLength: 40 }
      responses:
        '200':
          description: Paginated public rooms
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PaginatedRooms' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      operationId: createRoom
      x-fastify-config:
        rateLimit: { max: 20, timeWindow: '1 hour' }
      summary: Create a room; the caller becomes its owner
      tags: [rooms]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateRoomInput' }
      responses:
        '201':
          description: Room created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RoomResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
  /rooms/mine:
    get:
      operationId: listMyRooms
      summary: Rooms the caller is a member of (public and private), most recently active first
      tags: [rooms]
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Paginated rooms
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PaginatedRooms' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /rooms/{roomId}:
    parameters:
      - $ref: '#/components/parameters/RoomId'
    get:
      operationId: getRoom
      summary: Get a room
      description: Private rooms return 404 to non-members.
      tags: [rooms]
      responses:
        '200':
          description: Room
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RoomResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      operationId: updateRoom
      summary: Update a room (owner only)
      tags: [rooms]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateRoomInput' }
      responses:
        '200':
          description: Updated room
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RoomResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      operationId: deleteRoom
      summary: Delete a conversation (owner only)
      tags: [rooms]
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema: { $ref: '#/components/schemas/NullResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /rooms/{roomId}/join:
    parameters:
      - $ref: '#/components/parameters/RoomId'
    post:
      operationId: joinRoom
      summary: Join a room. Private rooms require a valid inviteCode. Idempotent.
      tags: [rooms]
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: '#/components/schemas/JoinRoomInput' }
      responses:
        '200':
          description: Joined (or already a member)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RoomResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /rooms/{roomId}/bots:
    parameters:
      - $ref: '#/components/parameters/RoomId'
    get:
      operationId: listRoomBots
      summary: Every available bot and whether it is seated in this room
      tags: [rooms]
      responses:
        '200':
          description: Bots with their seat in this room
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RoomBotList' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /rooms/{roomId}/bots/{userId}:
    parameters:
      - $ref: '#/components/parameters/RoomId'
      - $ref: '#/components/parameters/UserId'
    put:
      operationId: seatRoomBot
      summary: Add (seat) a bot in this room — owner only. Idempotent.
      tags: [rooms]
      responses:
        '200':
          description: Seated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RoomBotResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      operationId: kickRoomBot
      summary: Remove (kick) a bot from this room — owner only; it stops at once. Its past items stay. Idempotent.
      tags: [rooms]
      responses:
        '200':
          description: Kicked
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RoomBotResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /rooms/{roomId}/participants:
    parameters:
      - $ref: '#/components/parameters/RoomId'
    get:
      operationId: listRoomParticipants
      summary: Who is in this conversation — members and seated bots, with live presence
      description: |
        One roster for people and bots alike, so the client renders every participant
        the same way (doc/08 §2.3). `present` comes from deduplicated room presence;
        seated bots are always present. Private rooms return 404 to non-members.
      tags: [rooms]
      responses:
        '200':
          description: Participants, present first
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ParticipantList' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /rooms/{roomId}/members/me:
    parameters:
      - $ref: '#/components/parameters/RoomId'
    delete:
      operationId: leaveRoom
      summary: Leave a room. Owners cannot leave their own room.
      tags: [rooms]
      responses:
        '200':
          description: Left
          content:
            application/json:
              schema: { $ref: '#/components/schemas/NullResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /rooms/{roomId}/invite-code:
    parameters:
      - $ref: '#/components/parameters/RoomId'
    post:
      operationId: rotateInviteCode
      summary: Issue a new invite code for a private room (owner only); the old code stops working
      tags: [rooms]
      responses:
        '200':
          description: Room with new inviteCode
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RoomResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }

  # ─── items ─────────────────────────────────────────────────────────────────
  /rooms/{roomId}/items:
    parameters:
      - $ref: '#/components/parameters/RoomId'
    get:
      operationId: listRoomItems
      summary: Items in a room, ordered by number
      description: |
        Flat list; build branches client-side from `parentId`.
        Pass `after` (an item number) to fetch only items newer than one you already have —
        recovers newly created items only, not updates or deletions.
        Use `order=desc` for the newest window, then follow nextCursor for older pages.
      tags: [items]
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
        - in: query
          name: order
          schema: { type: string, enum: [asc, desc], default: asc }
          description: Number ordering; desc starts with the newest items and pages backward.
        - in: query
          name: after
          schema: { type: integer, minimum: 0 }
          description: Only return items with number greater than this.
      responses:
        '200':
          description: Paginated items
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PaginatedItems' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
    post:
      operationId: sendMessage
      x-fastify-config:
        rateLimit: { max: 60, timeWindow: '1 minute' }
      summary: Send a new message to a room (top-level)
      description: |
        Requires membership (public rooms auto-join on first post). Must include text or
        at least one media id (400 `EMPTY_ITEM`). `mediaIds` must be uploads owned by the
        caller and not yet attached. Creates a new Message and an Item in the room.
      tags: [items]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SendMessageInput' }
      responses:
        '201':
          description: Item created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ItemResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Session token. Web clients send it as the httpOnly `token` cookie; native clients as a Bearer header.
  parameters:
    Cursor:
      in: query
      name: cursor
      schema: { type: string, maxLength: 512 }
      description: Opaque cursor returned by the previous page.
    Limit:
      in: query
      name: limit
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
    SearchQuery:
      in: query
      name: q
      schema: { type: string, minLength: 1, maxLength: 200 }
      description: Simple text search within this resource.
    RoomId:
      in: path
      name: roomId
      required: true
      schema: { type: string, minLength: 1, maxLength: 64 }
    UserId:
      in: path
      name: userId
      required: true
      schema: { type: string, minLength: 1, maxLength: 64 }
...
  responses:
    BadRequest:
      description: Invalid input
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Unauthorized:
      description: Not authenticated
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Forbidden:
      description: Not allowed
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    NotFound:
      description: Not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Conflict:
      description: Conflicts with current state
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    TooManyRequests:
      description: Rate limited
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error: { type: string }
        code: { type: string }
    NullResponse:
      type: object
      required: [data]
      properties:
        data: { type: 'null' }
    PaginatedMeta:
      type: object
      required: [hasMore, nextCursor]
      properties:
        changeCursor: { type: string, description: Durable room change sequence for item pages }
        hasMore: { type: boolean }
        nextCursor: { type: [string, 'null'] }

    # ─── users ───────────────────────────────────────────────────────────────
    UserKind:
      type: string
      enum: [human, bot]
      description: Data for data-layer rules (human-authored activity, doc/08 I6). Components render `tag`, never branch on this.
    User:
      type: object
      required: [id, email, isGuest, kind, tag, displayName, avatarUrl, createdAt]
      properties:
        id: { type: string }
        email: { type: [string, 'null'] }
        isGuest: { type: boolean }
        kind: { $ref: '#/components/schemas/UserKind' }
        tag: { type: [string, 'null'], description: 'Short disclosure suffix shown after the name, e.g. BOT; null for people' }
        displayName: { type: string }
        avatarUrl: { type: [string, 'null'] }
        createdAt: { type: string, format: date-time }
    UserResponse:
      type: object
      required: [data]
      properties:
        data: { $ref: '#/components/schemas/User' }
...
    UpdateMeInput:
      type: object
      additionalProperties: false
      properties:
        displayName: { type: string, minLength: 1, maxLength: 50 }
        avatarUrl: { type: [string, 'null'], maxLength: 512 }
    UploadAvatarInput:
      type: object
      required: [file]
      properties:
        file: { type: string, format: binary }
    Author:
      type: object
      required: [id, name, avatarUrl, kind, tag]
      properties:
        id: { type: string }
        name: { type: string }
        avatarUrl: { type: [string, 'null'] }
        kind: { $ref: '#/components/schemas/UserKind' }
        tag: { type: [string, 'null'], description: 'Short disclosure suffix shown after the name, e.g. BOT; null for people' }
    ParticipantRole:
      type: string
      enum: [owner, member, bot]
    Participant:
      type: object
      required: [user, role, present]
      properties:
        user: { $ref: '#/components/schemas/Author' }
        role: { $ref: '#/components/schemas/ParticipantRole' }
        present: { type: boolean, description: Connected to this room now (deduplicated; a refresh does not flip it) }
    AuthorList:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/Author' }
    AuthorResponse:
      type: object
      required: [data]
      properties:
        data: { $ref: '#/components/schemas/Author' }
    RoomBot:
      type: object
      required: [user, handle, kind, seated]
      properties:
        user: { $ref: '#/components/schemas/Author' }
        handle: { type: string }
        kind: { type: string, enum: [house, optional], description: 'house = seated by default; optional = only when added' }
        seated: { type: boolean }
    RoomBotList:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/RoomBot' }
    RoomBotResponse:
      type: object
      required: [data]
      properties:
        data: { $ref: '#/components/schemas/RoomBot' }
    ParticipantList:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/Participant' }

    # ─── rooms ───────────────────────────────────────────────────────────────
    RoomVisibility:
      type: string
      enum: [public, private]
    RoomRole:
      type: string
      enum: [owner, member]
    Room:
      type: object
      required:
        [id, number, title, description, thumbnail, topic, visibility, ownerId, itemCount, responseCount, durationMs, memberCount, lastActivityAt, createdAt, role, inviteCode]
      properties:
        id: { type: string }
        number: { type: integer, description: 'Human-facing room sequence ("CONVERSATION 018")' }
        title: { type: string }
        description: { type: string }
        thumbnail:
          description: Conversation image
          oneOf:
            - $ref: '#/components/schemas/Media'
            - type: 'null'
        topic: { type: [string, 'null'] }
        visibility: { $ref: '#/components/schemas/RoomVisibility' }
        ownerId: { type: string }
        itemCount: { type: integer }
        responseCount: { type: integer, description: Live items after the opening one }
        durationMs: { type: integer, description: Total audio/video time of the live items (ms) }
        memberCount: { type: integer }
        lastActivityAt: { type: string, format: date-time }
        lastResponseAt: { type: [string, 'null'], format: date-time, description: 'Newest live item after the opening one; null when there is none' }
        createdAt: { type: string, format: date-time }
        role:
          description: The caller's membership role, or null if not a member
          oneOf:
            - $ref: '#/components/schemas/RoomRole'
            - type: 'null'
        inviteCode:
          type: [string, 'null']
          description: Present only for members of a private room
    RoomResponse:
      type: object
      required: [data]
      properties:
        data: { $ref: '#/components/schemas/Room' }
    PaginatedRooms:
      type: object
      required: [data, meta]
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/Room' }
        meta: { $ref: '#/components/schemas/PaginatedMeta' }
    CreateRoomInput:
      type: object
      additionalProperties: false
      description: Every field is optional; a conversation may have no name, description or thumbnail.
      properties:
        title: { type: string, maxLength: 120 }
        description: { type: string, maxLength: 4000 }
        thumbnailId:
          type: string
          minLength: 1
          maxLength: 64
          description: The caller's own image, or a YouTube video they added. 400 INVALID_THUMBNAIL otherwise.
        topic: { type: string, minLength: 1, maxLength: 40 }
        visibility: { $ref: '#/components/schemas/RoomVisibility' }
    UpdateRoomInput:
      type: object
      additionalProperties: false
      properties:
        title: { type: string, maxLength: 120 }
        description: { type: string, maxLength: 4000 }
        thumbnailId: { type: string, minLength: 1, maxLength: 64 }
        topic: { type: [string, 'null'], minLength: 1, maxLength: 40 }
        visibility: { $ref: '#/components/schemas/RoomVisibility' }
    JoinRoomInput:
      type: object
      additionalProperties: false
      properties:
        inviteCode: { type: string, minLength: 1, maxLength: 32 }

    # ─── items ───────────────────────────────────────────────────────────────
...
    Media:
      type: object
      required: [id, type, source, url, mimeType, size, duration, name, poster, externalId, title, embeddable]
      properties:
        id: { type: string }
        type: { $ref: '#/components/schemas/MediaType' }
        source:
          type: string
          enum: [stored, youtube]
          description: stored = our upload/capture; youtube = external video (never downloaded)
        url: { type: string }
        mimeType: { type: string }
        size: { type: integer, description: Bytes }
        duration: { type: [number, 'null'], description: Seconds (audio/video) }
        name: { type: [string, 'null'] }
        poster: { type: [string, 'null'], description: Poster image URL (video) }
        externalId: { type: [string, 'null'], description: Canonical YouTube video id (source youtube) }
        title: { type: [string, 'null'], description: External title (source youtube) }
        embeddable:
          type: boolean
          description: Whether the media can be played inline. Non-embeddable YouTube shows a link card; its duration is null, so it cannot be anchored.
...
    Message:
      type: object
      description: Reusable content (author, text, media). Hydrated on every Item.
      required: [id, author, text, media, createdAt]
      properties:
        id: { type: string }
        author: { $ref: '#/components/schemas/Author' }
        text: { type: [string, 'null'], description: Null when the Item placement is deleted }
        media:
          type: array
          description: Empty when the Item placement is deleted
          items: { $ref: '#/components/schemas/Media' }
        createdAt: { type: string, format: date-time }
    Item:
      type: object
      description: |
        Placement of a Message inside one room. The same Message can be placed in many
        rooms (share); each placement has its own number, thread position and reactions.
      required: [id, roomId, messageId, number, parentId, anchorStartMs, chat, message, reactions, createdAt, deletedAt]
      properties:
        id: { type: string }
        roomId: { type: string }
        messageId: { type: string }
        number: { type: integer, description: 'Per-room sequence ("RE:003")' }
        parentId: { type: [string, 'null'] }
        anchorStartMs:
          type: [integer, 'null']
          minimum: 0
          description: Moment (ms) in the parent's media this reply anchors to; null when un-anchored. A point, not a range.
        chat:
          type: boolean
          description: Posted from the room chat. These stay in the chat and are not played on the stage.
        message: { $ref: '#/components/schemas/Message' }
        reactions:
          type: array
          items: { $ref: '#/components/schemas/ReactionSummary' }
        createdAt: { type: string, format: date-time }
        deletedAt: { type: [string, 'null'], format: date-time }
    ItemResponse:
      type: object
      required: [data]
      properties:
        data: { $ref: '#/components/schemas/Item' }
    PaginatedItems:
      type: object
      required: [data, meta]
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/Item' }
        meta: { $ref: '#/components/schemas/PaginatedMeta' }
    SendMessageInput:
      type: object
      additionalProperties: false
      properties:
        text: { type: string, maxLength: 4000 }
        mediaIds:
          type: array
          maxItems: 10
          uniqueItems: true
          items: { type: string, minLength: 1, maxLength: 64 }
        chat:
          type: boolean
          description: Room chat message. Not a stage broadcast, and not a reply.
...